+++ package.json
@@ -7,13 +7,16 @@
     "@emotion/styled": "^11.11.0",
     "@mui/joy": "^5.0.0-beta.1",
     "@mui/material": "^5.14.4",
+    "@react-oauth/google": "^0.11.1",
     "@reduxjs/toolkit": "^1.8.2",
     "@types/aria-query": "^5.0.1",
+    "@types/gapi.auth2": "^0.0.57",
     "@types/node": "^16.18.39",
     "@types/react": "^18.2.17",
     "@types/react-dom": "^18.2.7",
     "axios": "^0.27.2",
     "bootstrap-icons": "^1.10.5",
+    "gapi-script": "^1.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.32.0",
