+++ src/pages/not-found/Not-fount.tsx
@@ -1,0 +1,11 @@
+function PageNotFound() {
+  return (
+    <>
+      <h1>
+        404 Page not found
+      </h1>
+    </>
+  );
+}
+
+export default PageNotFound;
